+++ src/x_notify.c
@@ -1,9 +1,10 @@
 #include "x_dg.h"
+#include <stdint.h>
 
 void DwNotifyUndoButton(Widget wg,View w,int evt,void* obj,void* d) {
   if (w->app==NULL) return;
 
-  switch ((int)d) {
+  switch ((uintptr_t)d) {
     case 0:
       SetSensitiveEx(wg,!!GroupCount(w->app->undoStack));
       break;
@@ -27,7 +28,7 @@ void DwNotifyIfExists(Widget wg,View w,int evt,void* obj,void* d) {
 
   if (w->app==NULL) return;
 
-  switch((int)d) {
+  switch((uintptr_t)d) {
     case T_MARKING:
       SetSensitiveEx(wg,!IsEmptyGroup(w->app->mark));
       break;
@@ -98,7 +99,7 @@ void DwNotifyIfNotExists(Widget wg,View w,int evt,void* obj,void* d) {
 
   if (w->app==NULL) return;
 
-  switch((int)d) {
+  switch((uintptr_t)d) {
     case T_MARKING:
       SetSensitiveEx(wg,IsEmptyGroup(w->app->mark));
       break;
@@ -165,7 +166,7 @@ void DwNotifyIfNotExists(Widget wg,View w,int evt,void* obj,void* d) {
 void DwNotifyShowFlags(Widget wg,View w,int evt,void* obj,void* d) {
   if (w->app==NULL) return;
 
-  switch ((int)d) {
+  switch ((uintptr_t)d) {
     case 0:
       SetSensitiveEx(wg,!!(w->showFlags & SHW_STRETCH));
       break;
