+++ src/x_view.c
@@ -11,6 +11,7 @@
 \**********************************************************************/
 
 #include "x_dg.h"
+#include <stdint.h>
 
 #define XTBGS XmToggleButtonGetState
 #define XTBSS XmToggleButtonSetState
@@ -1662,7 +1663,7 @@ void DelDependentWidgetEx(View w,Widget wg,int bWidgetDestroyed) {
 
 void AddDWProc(Widget wg,XtPointer xtpV,XtPointer mask,XtPointer obj,
     XtPointer xtpProc,XtPointer userData) {
-  AddDependentWidget((View)xtpV,wg,(int)mask,(void*)obj,
+  AddDependentWidget((View)xtpV,wg,(uintptr_t)mask,(void*)obj,
     (DwNotifyProc)xtpProc,(void*)userData);
 }
